copied code for joystick over from pi4j test module, implementation doesnt work yet

+++ IP12-Prototype/src/main/java/ch/IP12/prototype/model/Moveable.java
@@ -9,10 +9,12 @@ import javafx.scene.canvas.GraphicsContext;
 public abstract class Moveable {
     protected double x;
     protected double y;
+
     protected final int speed;
     protected final double length;
     protected final double height;
-    protected int direction;
+
+    protected double direction;
 
     protected final double maxX;
     protected final double maxY;
